Push delete config and row
