+++ src/tests/push-delete-branch-main/expected-stderr
@@ -1,1 +1,1 @@
-Error: cannot perform the "push" operation:invalid action "delete branch main", default branch can never be deleted
+Error: cannot perform the "push" operation:invalid action delete branch "main", default branch can never be deleted
