+++ src/tests/push-delete-config-dev/expected-code
@@ -1,0 +1,1 @@
+1
