+++ src/tests/push-delete-config-dev/expected-stderr
@@ -1,0 +1,1 @@
+Error: cannot perform the "push" operation:invalid action delete config "foo/extractor/keboola.ex-db-mysql/with-rows", cannot delete config from dev branch
