+++ src/tests/push-delete-config-dev/expected-stdout
@@ -1,0 +1,2 @@
+Plan for "push" operation:
+×  C foo/extractor/keboola.ex-db-mysql/with-rows
