+++ src/tests/push-delete-config-dev/initial-state.json
@@ -1,0 +1,20 @@
+{
+  "allBranchesConfigs": [],
+  "branches": [
+    {
+      "branch": {
+        "name": "Main",
+        "isDefault": true
+      }
+    },
+    {
+      "branch": {
+        "name": "Foo",
+        "isDefault": false
+      },
+      "configs": [
+        "with-rows"
+      ]
+    }
+  ]
+}
