+++ src/tests/push-delete-config-main/expected-code
@@ -1,0 +1,1 @@
+0
