+++ src/tests/push-delete-config-main/expected-stdout
@@ -1,0 +1,3 @@
+Plan for "push" operation:
+×  C main/extractor/keboola.ex-db-mysql/with-rows
+Push done.
