+++ src/tests/push-delete-config-main/initial-state.json
@@ -1,0 +1,14 @@
+{
+  "allBranchesConfigs": [],
+  "branches": [
+    {
+      "branch": {
+        "name": "Main",
+        "isDefault": true
+      },
+      "configs": [
+        "with-rows"
+      ]
+    }
+  ]
+}
