+++ src/tests/push-delete-config-row-dev/args
@@ -1,0 +1,1 @@
+push --storage-api-host %%TEST_KBC_STORAGE_API_HOST%% --storage-api-token %%TEST_KBC_STORAGE_API_TOKEN_MASTER%%
