+++ src/tests/push-delete-config-row-dev/expected-code
@@ -1,0 +1,1 @@
+1
