+++ src/tests/push-delete-config-row-dev/expected-stderr
@@ -1,0 +1,3 @@
+Error: cannot perform the "push" operation:
+- invalid action delete config row "foo/extractor/keboola.ex-db-mysql/with-rows/rows/test-view", cannot delete config row from dev branch
+- invalid action delete config row "foo/extractor/keboola.ex-db-mysql/with-rows/rows/users", cannot delete config row from dev branch
