+++ src/tests/push-delete-config-row-dev/in/foo/extractor/keboola.ex-db-mysql/with-rows/config.json
@@ -1,0 +1,7 @@
+{
+  "parameters": {
+    "db": {
+      "host": "mysql.example.com"
+    }
+  }
+}
