+++ src/tests/push-delete-config-row-dev/in/foo/extractor/keboola.ex-db-mysql/with-rows/rows/disabled/config.json
@@ -1,0 +1,5 @@
+{
+  "parameters": {
+    "incremental": false
+  }
+}
