+++ src/tests/push-delete-config-row-dev/in/foo/meta.json
@@ -1,0 +1,5 @@
+{
+  "name": "Foo",
+  "description": "",
+  "isDefault": false
+}
