+++ src/tests/push-delete-config-row-dev/in/main/meta.json
@@ -1,0 +1,5 @@
+{
+  "name": "Main",
+  "description": "",
+  "isDefault": true
+}
