+++ src/tests/push-delete-config-row-main/expected-code
@@ -1,0 +1,1 @@
+0
