+++ src/tests/push-delete-config-row-main/expected-stdout
@@ -1,0 +1,4 @@
+Plan for "push" operation:
+×  R main/extractor/keboola.ex-db-mysql/with-rows/rows/test-view
+×  R main/extractor/keboola.ex-db-mysql/with-rows/rows/users
+Push done.
