+++ src/tests/push-delete-config-row-main/in/main/extractor/keboola.ex-db-mysql/with-rows/config.json
@@ -1,0 +1,7 @@
+{
+  "parameters": {
+    "db": {
+      "host": "mysql.example.com"
+    }
+  }
+}
