+++ src/tests/push-delete-config-row-main/in/main/extractor/keboola.ex-db-mysql/with-rows/meta.json
@@ -1,0 +1,4 @@
+{
+  "name": "with-rows",
+  "description": "test fixture"
+}
