+++ src/tests/push-delete-config-row-main/in/main/extractor/keboola.ex-db-mysql/with-rows/rows/disabled/config.json
@@ -1,0 +1,5 @@
+{
+  "parameters": {
+    "incremental": false
+  }
+}
