+++ src/tests/push-delete-config-row-main/in/main/extractor/keboola.ex-db-mysql/with-rows/rows/disabled/meta.json
@@ -1,0 +1,5 @@
+{
+  "name": "disabled",
+  "description": "test fixture",
+  "isDisabled": true
+}
